chore: Add argparse for directory variable, outsorcing file class to a own file and improve the main.py
